+++ StagingController.cs
@@ -1,3 +1,5 @@
+using System.Reflection;
+
 using KRPC.Service.Attributes;
 
 namespace KRPC.MechJeb {
@@ -11,6 +13,8 @@ namespace KRPC.MechJeb {
 		private readonly object clampAutoStageThrustPct;
 		private readonly object fairingMaxAerothermalFlux;
 
+		private readonly FieldInfo autostagingOnce;
+
 		public StagingController() : base("StagingController") {
 			this.autostagePreDelay = this.type.GetField("autostagePreDelay").GetValue(this.instance);
 			this.autostagePostDelay = this.type.GetField("autostagePostDelay").GetValue(this.instance);
@@ -19,6 +23,8 @@ namespace KRPC.MechJeb {
 			this.fairingMinAltitude = this.type.GetField("fairingMinAltitude").GetValue(this.instance);
 			this.clampAutoStageThrustPct = this.type.GetField("clampAutoStageThrustPct").GetValue(this.instance);
 			this.fairingMaxAerothermalFlux = this.type.GetField("fairingMaxAerothermalFlux").GetValue(this.instance);
+
+			this.autostagingOnce = this.type.GetField("autostagingOnce");
 		}
 
 		[KRPCProperty]
@@ -64,6 +70,12 @@ namespace KRPC.MechJeb {
 		public double FairingMaxAerothermalFlux {
 			get => EditableVariables.GetDouble(this.fairingMaxAerothermalFlux);
 			set => EditableVariables.SetDouble(this.fairingMaxAerothermalFlux, value);
+		}
+
+		[KRPCProperty]
+		public bool AutostagingOnce {
+			get => (bool)this.autostagingOnce.GetValue(this.instance);
+			set => this.autostagingOnce.SetValue(this.instance, value);
 		}
 	}
 }